like post Done

+++ backend/src/main/java/com/ecommerce/backend/repository/UserLikePostRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserLikePostRepository extends JpaRepository<Report, Long> {
     @Query(value = "SELECT COUNT(pl.user_id) FROM user_post_like pl WHERE post_id = :postId GROUP BY pl.post_id", nativeQuery = true)
-    Integer countUsersLike(Long postId);
+    Long countUsersLike(Long postId);
 
 }
